Fix and modernize useGroups/useGroup/useDiscoveryFeed/GroupDetailPage test suites
- useGroups.test.ts and useGroup.test.ts never rendered inside a
  QueryClientProvider, so every test failed immediately with "No
  QueryClient set" once the hooks were on React Query -- they were
  apparently written/left mid-migration. Renamed to .tsx and wrapped
  each render in a fresh QueryClient. useGroup.test.tsx also drops
  assertions that no longer match the hook's real behavior (it never
  threw a "Group not found." message) and adds a test proving
  useGroup() and usePrefetchGroup() share one cache entry (the actual
  "no duplicate network calls" acceptance criterion for hover-prefetch
  -> detail page navigation).
- useGroups.ts: applyFilters() silently ignored minCycleDuration/
  maxCycleDuration even though hasActiveFilters() and the GroupFilters
  UI both treat them as real filters -- fixed so the cycle-duration
  inputs actually filter results, and added tests.
- useDiscoveryFeed.test.ts (BrowseGroupsPage's data hook): same missing
  QueryClientProvider issue; renamed to .tsx, wrapped, and added a test
  that two mounts on one QueryClient reuse the cached group list instead
  of double-fetching.
- GroupDetailPage.snapshot.test.tsx mocked a `fetchDetailedGroup` export
  that never existed in utils/groupApi.ts, plus called vi.mock() inside
  test bodies referencing not-yet-hoisted variables -- it never actually
  ran. Rewritten against the real fetchGroup/useGroup() data path with a
  QueryClientProvider wrapper and DetailedGroup-shaped fixtures; snapshots
  regenerated.

Part of #1263.
