+++ frontend/src/test/GroupDetailPage.snapshot.test.tsx
@@ -1,12 +1,20 @@
 /**
  * Snapshot tests for GroupDetailPage
  *
- * Captures rendering state to detect unintended UI regressions
+ * Captures rendering state to detect unintended UI regressions. Rewritten
+ * to match GroupDetailPage's current data source: the shared useGroup()
+ * React Query hook (queryKeys.groups.detail), not the page's old internal
+ * mock-data generator. Mocks `utils/groupApi`'s real `fetchGroup` export
+ * (the previous version of this file mocked a `fetchDetailedGroup` export
+ * that was never part of groupApi.ts, so it never actually exercised the
+ * page).
  */
 import { render } from '@testing-library/react';
 import { describe, it, expect, vi, beforeEach } from 'vitest';
-import { MemoryRouter, Route, Routes } from 'react-router-dom';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import GroupDetailPage from '../pages/GroupDetailPage';
+import * as groupApi from '../utils/groupApi';
+import type { DetailedGroup } from '../utils/groupApi';
 
 // ── Mock dependencies ─────────────────────────────────────────────────────────
 
@@ -21,48 +29,71 @@ vi.mock('../ui', () => ({
 }));
 
 vi.mock('../hooks/useWallet', () => ({
-  useWallet: () => ({ address: 'GABC123XYZ', isConnected: true }),
+  useWallet: () => ({ activeAddress: 'GABC123XYZ' }),
 }));
 
 vi.mock('../routing/useNavigation', () => ({
-  useNavigation: () => ({ navigateTo: vi.fn() }),
+  useNavigation: () => ({ params: { groupId: '123' } }),
 }));
 
-// Mock fixture data
-const mockGroupDetail = {
+// Heavy subtrees mocked out — this suite verifies GroupDetailPage's own
+// loading/error/data rendering via useGroup(), not these components'
+// internals (each has its own test coverage elsewhere).
+vi.mock('../components/ContributionFlow', () => ({
+  ContributionFlow: () => <div data-testid="contribution-flow" />,
+}));
+vi.mock('../components/InsurancePanel', () => ({
+  InsurancePanel: () => <div data-testid="insurance-panel" />,
+}));
+vi.mock('../components/GroupReportExportButton', () => ({
+  GroupReportExportButton: () => <button type="button">Export report</button>,
+}));
+
+vi.mock('../utils/groupApi', async () => {
+  const actual = await vi.importActual<typeof import('../utils/groupApi')>('../utils/groupApi');
+  return { ...actual, fetchGroup: vi.fn() };
+});
+
+const mockFetchGroup = vi.mocked(groupApi.fetchGroup);
+
+// Mock fixture data — shape matches DetailedGroup (the real return type of
+// fetchGroup / useGroup()'s `group` field).
+const mockGroupDetail: DetailedGroup = {
   id: '123',
   name: 'Alpha Savings Circle',
   description: 'Monthly savings pool for community members',
   status: 'active',
+  memberCount: 5,
   contributionAmount: 100,
   currency: 'XLM',
-  cycleNumber: 3,
-  cycleDuration: 604800,
-  maxMembers: 10,
-  memberCount: 7,
   createdAt: new Date('2024-01-15'),
-  creator: 'GCREATOR123',
+  totalMembers: 5,
+  targetAmount: 500,
+  currentAmount: 300,
+  contributionFrequency: 'monthly',
   members: [
-    { address: 'GMEMBER1', joinedAt: new Date('2024-01-15'), payoutPosition: 1, hasReceivedPayout: true },
-    { address: 'GMEMBER2', joinedAt: new Date('2024-01-16'), payoutPosition: 2, hasReceivedPayout: true },
-    { address: 'GMEMBER3', joinedAt: new Date('2024-01-17'), payoutPosition: 3, hasReceivedPayout: true },
-    { address: 'GABC123XYZ', joinedAt: new Date('2024-01-18'), payoutPosition: 4, hasReceivedPayout: false },
-    { address: 'GMEMBER5', joinedAt: new Date('2024-01-19'), payoutPosition: 5, hasReceivedPayout: false },
+    { id: '1', address: 'GMEMBER1', joinedAt: new Date('2024-01-15'), totalContributions: 300, isActive: true },
+    { id: '2', address: 'GMEMBER2', joinedAt: new Date('2024-01-16'), totalContributions: 300, isActive: true },
+    { id: '3', address: 'GABC123XYZ', joinedAt: new Date('2024-01-18'), totalContributions: 200, isActive: true },
   ],
   contributions: [
-    { memberId: 'GMEMBER1', cycle: 3, amount: 100, timestamp: new Date('2024-06-01') },
-    { memberId: 'GMEMBER2', cycle: 3, amount: 100, timestamp: new Date('2024-06-01') },
-    { memberId: 'GABC123XYZ', cycle: 3, amount: 100, timestamp: new Date('2024-06-02') },
+    { id: 'c1', memberId: '1', memberName: 'GMEMBER1', amount: 100, timestamp: new Date('2024-06-01'), transactionHash: 'tx1', status: 'completed' },
   ],
-  payouts: [
-    { cycle: 1, recipient: 'GMEMBER1', amount: 700, timestamp: new Date('2024-03-01') },
-    { cycle: 2, recipient: 'GMEMBER2', amount: 700, timestamp: new Date('2024-04-01') },
+  cycles: [
+    { cycleNumber: 1, startDate: new Date('2024-01-01'), endDate: new Date('2024-01-31'), targetAmount: 500, currentAmount: 500, status: 'completed' },
+    { cycleNumber: 2, startDate: new Date('2024-02-01'), endDate: new Date('2024-02-28'), targetAmount: 500, currentAmount: 300, status: 'active' },
   ],
+  currentCycle: { cycleNumber: 2, startDate: new Date('2024-02-01'), endDate: new Date('2024-02-28'), targetAmount: 500, currentAmount: 300, status: 'active' },
 };
 
-vi.mock('../utils/groupApi', () => ({
-  fetchDetailedGroup: vi.fn().mockResolvedValue(mockGroupDetail),
-}));
+function renderPage() {
+  const queryClient = new QueryClient({ defaultOptions: { queries: { retry: false } } });
+  return render(
+    <QueryClientProvider client={queryClient}>
+      <GroupDetailPage />
+    </QueryClientProvider>
+  );
+}
 
 // ── Snapshot tests ────────────────────────────────────────────────────────────
 
@@ -72,71 +103,40 @@ describe('GroupDetailPage snapshot tests', () => {
   });
 
   it('matches snapshot with full group data', async () => {
-    const { container, findByText } = render(
-      <MemoryRouter initialEntries={['/groups/123']}>
-        <Routes>
-          <Route path="/groups/:groupId" element={<GroupDetailPage />} />
-        </Routes>
-      </MemoryRouter>
-    );
+    mockFetchGroup.mockResolvedValue(mockGroupDetail);
+    const { container, findAllByText } = renderPage();
 
-    // Wait for async data to load
-    await findByText('Alpha Savings Circle');
+    // The group name renders both as the AppLayout page title and as the
+    // in-content heading, so there are two matching elements.
+    await findAllByText('Alpha Savings Circle');
     expect(container).toMatchSnapshot();
   });
 
   it('matches snapshot with loading state', () => {
-    vi.mock('../utils/groupApi', () => ({
-      fetchDetailedGroup: vi.fn().mockReturnValue(new Promise(() => {})),
-    }));
-
-    const { container } = render(
-      <MemoryRouter initialEntries={['/groups/123']}>
-        <Routes>
-          <Route path="/groups/:groupId" element={<GroupDetailPage />} />
-        </Routes>
-      </MemoryRouter>
-    );
+    mockFetchGroup.mockReturnValue(new Promise(() => {}));
+    const { container } = renderPage();
     expect(container).toMatchSnapshot();
   });
 
   it('matches snapshot with error state', async () => {
-    vi.mock('../utils/groupApi', () => ({
-      fetchDetailedGroup: vi.fn().mockRejectedValue(new Error('Group not found')),
-    }));
-
-    const { container, findByText } = render(
-      <MemoryRouter initialEntries={['/groups/123']}>
-        <Routes>
-          <Route path="/groups/:groupId" element={<GroupDetailPage />} />
-        </Routes>
-      </MemoryRouter>
-    );
+    mockFetchGroup.mockRejectedValue(new Error('Group not found'));
+    const { container, findByText } = renderPage();
 
     await findByText(/error/i, {}, { timeout: 3000 }).catch(() => {});
     expect(container).toMatchSnapshot();
   });
 
   it('matches snapshot with completed group', async () => {
-    const completedGroup = {
+    const completedGroup: DetailedGroup = {
       ...mockGroupDetail,
       status: 'completed',
-      members: mockGroupDetail.members.map(m => ({ ...m, hasReceivedPayout: true })),
+      members: mockGroupDetail.members.map((m) => ({ ...m, isActive: false })),
     };
+    mockFetchGroup.mockResolvedValue(completedGroup);
 
-    vi.mock('../utils/groupApi', () => ({
-      fetchDetailedGroup: vi.fn().mockResolvedValue(completedGroup),
-    }));
+    const { container, findAllByText } = renderPage();
 
-    const { container, findByText } = render(
-      <MemoryRouter initialEntries={['/groups/123']}>
-        <Routes>
-          <Route path="/groups/:groupId" element={<GroupDetailPage />} />
-        </Routes>
-      </MemoryRouter>
-    );
-
-    await findByText('Alpha Savings Circle');
+    await findAllByText('Alpha Savings Circle');
     expect(container).toMatchSnapshot();
   });
 });
